refactor: organize runtime modules by lifecycle domain
