+++ lib/jido_harness/session/adapter.ex
