+++ lib/jido_harness/session/info.ex
