+++ lib/jido_harness/session/interaction_capabilities.ex
