+++ lib/jido_harness/session/transport_spec.ex
